Refactor ThunderSparksHighlight to be in fishing
